Refactor internal Spriteset object
spriteset.c contained some of the oldest code in the engine.  It was one
of the first parts of the engine I developed and the code has barely
been touched since then.  I figured now was a good time to overhaul it,
as part of implementing Sphere 1.x-compatible spriteset handling.

There might still be bugs in this implementation; more testing will be
needed before this makes it into a release.

+++ src/shared/vector.h
@@ -12,20 +12,20 @@ struct iter
 {
 	vector_t* vector;
 	void*     ptr;
-	ptrdiff_t index;
+	int       index;
 } iter_t;
 
 vector_t* vector_new    (size_t pitch);
 vector_t* vector_dup    (const vector_t* vector);
 void      vector_free   (vector_t* vector);
 void      vector_clear  (vector_t* vector);
-size_t    vector_len    (const vector_t* vector);
+int       vector_len    (const vector_t* vector);
 bool      vector_push   (vector_t* vector, const void* in_object);
-void      vector_remove (vector_t* vector, size_t index);
+void      vector_remove (vector_t* vector, int index);
 vector_t* vector_sort   (vector_t* vector, int(*comparer)(const void* in_a, const void* in_b));
 
-void*  vector_get  (const vector_t* vector, size_t index);
-void   vector_set  (vector_t* vector, size_t index, const void* in_object);
+void*  vector_get  (const vector_t* vector, int index);
+void   vector_set  (vector_t* vector, int index, const void* in_object);
 iter_t vector_enum (vector_t* vector);
 void*  vector_next (iter_t* inout_iter);
 void   iter_remove (iter_t* iter);
